+++ src/components/form/task/schemaTodo.tsx
@@ -2,7 +2,7 @@ import {z} from 'zod'
 
 const schemaToDo = z.object({
   time: z.string(),
-  title: z.string(),
+  title: z.string().min(4, {message: 'Sua tarefa não pode ficar sem título'}),
   description: z.string(),
 })
